Remove contract classes, keep viewStates
