--- WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/AnalyticsContract.kt
@@ -1,11 +1,0 @@
-package com.woocommerce.android.ui.analytics
-
-import com.woocommerce.android.ui.analytics.daterangeselector.AnalyticsDateRangeSelectorContract.AnalyticsDateRangeSelectorViewState
-
-class AnalyticsContract {
-    data class AnalyticsState(val analyticsDateRangeSelectorState: AnalyticsDateRangeSelectorViewState)
-
-    sealed class AnalyticsEvent
-
-    sealed class AnalyticsEffect
-}
