+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/AnalyticsViewState.kt
@@ -1,0 +1,5 @@
+package com.woocommerce.android.ui.analytics
+
+import com.woocommerce.android.ui.analytics.daterangeselector.AnalyticsDateRangeSelectorViewState
+
+data class AnalyticsViewState(val analyticsDateRangeSelectorState: AnalyticsDateRangeSelectorViewState)
