+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/analytics/daterangeselector/AnalyticsDateRangeSelectorViewState.kt
@@ -1,0 +1,8 @@
+package com.woocommerce.android.ui.analytics.daterangeselector
+
+data class AnalyticsDateRangeSelectorViewState(
+    val toDatePeriod: String,
+    val fromDatePeriod: String,
+    val availableRangeDates: List<String>,
+    val selectedPeriod: String
+)
